cookstyle fixes

+++ recipes/cis-5-system-services.rb
@@ -9,39 +9,37 @@
 if node['cis']['level']['1'] == 'yes'
   # apply level 1 items
   node['cis-5']['level1']['services'].each do |i|
-    if ::Win32::Service.exists?(i)
-      # check if the service actually exists first so we don't generate errors
-      Chef::Log.warn("Checking if service #{i} has any dependencies")
-      powershell_script 'DependencyCheck' do
-        code <<-EOH
-          Start-Transcript -Verbose
-          $serviceList = (Get-Service -Name #{i} | Select DependentServices).DependentServices
-            if( ($serviceList.count) -gt 0) {
-              Write-Warning "Dependent services have been found, going to brute force stop them..."
-              foreach($item in $serviceList){
-                Write-Warning "Attempting to stop service $($item.Name)"
-                Get-Service $item.Name | Stop-Service -Force -Verbose
-                }
+    next unless ::Win32::Service.exists?
+    # check if the service actually exists first so we don't generate errors
+    Chef::Log.warn("Checking if service #{i} has any dependencies")
+    powershell_script 'DependencyCheck' do
+      code <<-EOH
+        Start-Transcript -Verbose
+        $serviceList = (Get-Service -Name #{i} | Select DependentServices).DependentServices
+          if( ($serviceList.count) -gt 0) {
+            Write-Warning "Dependent services have been found, going to brute force stop them..."
+            foreach($item in $serviceList){
+              Write-Warning "Attempting to stop service $($item.Name)"
+              Get-Service $item.Name | Stop-Service -Force -Verbose
               }
-            Stop-Transcript -Verbose
-          EOH
-      end
-      windows_service i do
-        action [:stop, :disable]
-      end
+            }
+          Stop-Transcript -Verbose
+        EOH
+    end
+    windows_service i do
+      action [:stop, :disable]
     end
   end
 end
 
 # if CIS Level 2 is set to yes, then we apply the items
 if node['cis']['level']['2'] == 'yes'
-    # apply level 2 items
-    node['cis-5']['level1']['services'].each do |i|
-        if ::Win32::Service.exists?(i)
-            # check if the service actually exists first so we don't generate errors
-            windows_service i do
-                action [:stop, :disable]
-            end
-        end
+  # apply level 2 items
+  node['cis-5']['level1']['services'].each do |i|
+    next unless ::Win32::Service.exists?
+    # check if the service actually exists first so we don't generate errors
+    windows_service i do
+      action [:stop, :disable]
     end
+  end
 end
